+++ contrib/python/clickhouse-connect/clickhouse_connect/driver/dataconv.py
@@ -28,6 +28,9 @@ def read_ipv4_col(source: ByteSource, num_rows: int):
 
 def read_datetime_col(source: ByteSource, num_rows: int, tz_info: Optional[tzinfo]):
     src_array = source.read_array('I', num_rows)
+    if tz_info is None:
+        fts = datetime.utcfromtimestamp
+        return [fts(ts) for ts in src_array]
     fts = datetime.fromtimestamp
     return [fts(ts, tz_info) for ts in src_array]