+++ ydb/core/fq/libs/row_dispatcher/common.h
@@ -1,9 +1,0 @@
-#pragma once
-
-#include <util/generic/fwd.h>
-
-namespace NFq {
-
-TString CleanupCounterValueString(const TString& value);
-
-} // namespace NFq